Fixed MSVC/gcc conformance for Vector

+++ src/core/engine.cpp
@@ -59,7 +59,7 @@ bool Engine::initialize() {
 	//Test code
 	Vec2f a(1, 2);
 	Vec2f b(2.0, -1.7);
-	Vec2i c = a + b;
+	Vec2d c = a + b;
 
 	Log::info("", "%d, %d, %s", c.x, c.y, typeid(Math::mod(2.0, 1.0f)).name());
